+++ crates/toolchain/src/lib.rs
@@ -3,9 +3,9 @@ mod spec;
 pub use moon_config::{UnresolvedVersionSpec, VersionSpec};
 pub use spec::*;
 
+use indexmap::IndexSet;
 use moon_common::{Id, IdExt};
 use moon_env_var::{GlobalEnvBag, as_bool};
-use rustc_hash::FxHashSet;
 
 pub fn is_using_global_toolchains(bag: &GlobalEnvBag) -> bool {
     bag.get_as("MOON_TOOLCHAIN_FORCE_GLOBALS", as_bool)
@@ -57,7 +57,7 @@ pub fn filter_and_resolve_toolchain_ids(
     in_list: Vec<Id>,
     fallback_system: bool,
 ) -> Vec<Id> {
-    let mut out_list = FxHashSet::default();
+    let mut out_list = IndexSet::<Id>::default();
 
     for id in in_list {
         if id == "system" {